+++ front/web/src/App.tsx
@@ -3,18 +3,18 @@ import { GlobalStyles } from './styles/GlobalStyles';
 import PostChoice from './pages/PostChoice';
 import LoginPage from './pages/LoginPage';
 import PersonaChoice from './pages/PersonaChoicePage';
-import MainLayout from './components/commons/MainLayout';
+import TitleLayout from './components/commons/TitleLayout';
 
 const App = () => {
   return (
-    <MainLayout>
+    <TitleLayout>
       <GlobalStyles />
       <Routes>
         <Route path='/' element={<PostChoice/>} />
         <Route path='/login' element={<LoginPage />} />
         <Route path='/persona' element={<PersonaChoice />} />
       </Routes>
-    </MainLayout>
+    </TitleLayout>
   );
 }
